% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_report.R
\name{aa_anomaly_report}
\alias{aa_anomaly_report}
\title{Anomaly Report}
\usage{
aa_anomaly_report(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  date_range = c(Sys.Date() - 31, Sys.Date() - 1),
  metrics,
  pages = 0,
  granularity = "day",
  sort = "asc",
  anomalyDetection = TRUE
)
}
\arguments{
\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{rsid}{Adobe report number}

\item{date_range}{A two length vector of start and end Date objects (default set to show last 30 days)}

\item{metrics}{Metric to send}

\item{pages}{number of report pages}

\item{granularity}{use either minute, hour, day, week, month, quarter, year}

\item{sort}{either by 'desc' or 'asc' order}

\item{anomalyDetection}{logical statement for including anomaly. Default is TRUE}
}
\description{
Get an anomaly report for one or more metrics
}
