% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_data.R
\name{aa_call_data}
\alias{aa_call_data}
\title{Gets the data from Adobe Analytics API v2}
\usage{
aa_call_data(
  req_path,
  body = NULL,
  company_id = Sys.getenv("AA_COMPANY_ID"),
  client_id = Sys.getenv("AA_CLIENT_ID"),
  client_secret = Sys.getenv("AA_CLIENT_SECRET"),
  debug = FALSE
)
}
\arguments{
\item{req_path}{The endpoint for that particular report}

\item{body}{An R list that will be parsed to JSON}

\item{company_id}{Set in environment args, or pass directly here}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}

\item{debug}{Set this to TRUE to see the information about the api calls as they happen.}
}
\description{
This gives a raw call to the API, but it is intended other functions call this one
}
\examples{

\dontrun{

aa_call_data("reports/ranked",
            body = list(..etc..),
            company_id = "blah")

}

}
