% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adding_aa_credentials.R
\name{aa_creds}
\alias{aa_creds}
\title{Setting credentials to the environment}
\usage{
aa_creds()
}
\description{
This function will enable you to set the 4 essential arguments the API requires for OAuth and making calls
to specific report suites.
This is optional. Setting global variables or utilizing the arguments in the function call are the
two other options.
}
