% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{aa_get_data}
\alias{aa_get_data}
\title{Gets the data from Adobe Analytics API v2}
\usage{
aa_get_data(
  req_path,
  body = NULL,
  company_id = Sys.getenv("AA_COMPANY_ID"),
  client_id = Sys.getenv("AA_CLIENT_ID"),
  client_secret = Sys.getenv("AA_CLIENT_SECRET")
)
}
\arguments{
\item{req_path}{The endpoint for that particular report}

\item{body}{An R list that will be parsed to JSON}

\item{company_id}{Set in environment args, or pass directly here}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}
}
\description{
This gives a raw call to the API, but it is intended other functions call this one
}
\examples{

\dontrun{

aa_get_data("reports/ranked",
            body = list(..etc..),
            company_id = "blah")

}

}
