% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{aa_get_metrics}
\alias{aa_get_metrics}
\title{Get list of dimensions}
\usage{
aa_get_metrics(rsid = Sys.getenv("AA_RSID"), locale = "en_US",
  segmentable = "false", expansion = c("tags", "allowedForReporting",
  "categories"))
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{- boolean - default TRUE}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}
}
\description{
Get list of dimensions
}
