% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments.R
\name{aa_get_segments}
\alias{aa_get_segments}
\title{Get list of segments}
\usage{
aa_get_segments(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsids = NA,
  segmentFilter = NA,
  locale = "en_US",
  name = NA,
  tagNames = NA,
  filterByPublishedSegments = "all",
  limit = 10,
  page = 0,
  sortDirection = "ASC",
  sortProperty = "id",
  expansion = NA,
  includeType = "all"
)
}
\arguments{
\item{rsids}{Filter list to only include segments tied to specified RSID list (comma-delimited)}

\item{segmentFilter}{Filter list to only include suites in this RSID list (comma-delimited)}

\item{locale}{Locale (en_US default)}

\item{name}{Filter list to only include segments that contains the Name}

\item{tagNames}{Filter list to only include segments that contains one of the tags}

\item{filterByPublishedSegments}{Filter list to only include segments where the published field is set to one of the allowable values (all *default*, true, false)}

\item{limit}{Number of results per page (10 - *default*)}

\item{page}{Page number (base 0 - first page is "0")}

\item{sortDirection}{Sort direction (ASC *default* or DESC)}

\item{sortProperty}{Property to sort by (name, modified_date, id *default* is currently allowed)}

\item{expansion}{Comma-delimited list of additional segment metadata fields to include on response. (Options--reportSuiteName, ownerFullName, modified,tags, compatibility, definition, publishingStatus, definitionLastModified, categories)}

\item{includeType}{Include additional segments not owned by user. The "all" option takes precedence over "shared" (shared, all, templates)}
}
\description{
Retrieve All Segments
}
