% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_report.R
\name{aa_ranked_report}
\alias{aa_ranked_report}
\title{Get a ranked report}
\usage{
aa_ranked_report(
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  date_range,
  metrics,
  dimensions,
  top = 50,
  metricSort = FALSE,
  dimensionSort = "asc",
  company_id = Sys.getenv("AA_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report suite id number. Taken from the global environment by default if not provided.}

\item{date_range}{A two length vector of start and end Date objects}

\item{metrics}{Metric to send}

\item{dimensions}{Dimension to send}

\item{top}{How many rows. Defualt is set to 50}

\item{metricSort}{Presorts the table by metrics. Values are either 'asc' or 'desc'.}

\item{dimensionSort}{Presorts the table by dimension.  Not as helpful as metricSort but still has some use.  If metricSort is 'False' the dimension sort will be used.}
}
\value{
Data Frame
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Only runs a single dimension with as many metrics as you want.
}
