% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked_time_report.R
\name{aa_rankedtime_report}
\alias{aa_rankedtime_report}
\title{Get a timeseries reports}
\usage{
aa_rankedtime_report(date_range, metrics, dimension, top = 0,
  pages = 0, granularity = "day", metricFilters = NULL,
  sort = "asc", rsid = Sys.getenv("AA_RSID"))
}
\arguments{
\item{date_range}{A two length vector of start and end Date objects}

\item{metrics}{Metric to send}

\item{rsid}{Adobe report number}

\item{dimensions}{Dimension to send}

\item{num_rows}{How many rows}
}
\description{
Get a timeseries reports
}
