% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_report.R
\name{aa_timeseries_report}
\alias{aa_timeseries_report}
\title{Get a timeseries reports}
\usage{
aa_timeseries_report(
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  company_id = Sys.getenv("AA_COMPANY_ID"),
  date_range,
  metrics,
  top = 50,
  pages = 0,
  granularity = "day",
  metricFilters = NULL,
  sort = "asc",
  anomalyDetection = TRUE
)
}
\arguments{
\item{rsid}{Adobe report number}

\item{date_range}{A two length vector of start and end Date objects}

\item{metrics}{Metric to send}

\item{top}{the number of rows to pull}

\item{granularity}{the breakdown of the date - hour, day, month, quarter, year}

\item{metricFilters}{the filters to breakdown the metrics by (still in dev)}

\item{sort}{the order of date either asc or desc (note: is it good that there is no sorting by metrics?)}

\item{anomalyDetection}{Anomaly detection logic}
}
\description{
Get a timeseries reports
}
