% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{aa_token}
\alias{aa_token}
\title{OAuth2 Token for Adobe Analytics}
\usage{
aa_token(
  client_id = Sys.getenv("AA_CLIENT_ID"),
  client_secret = Sys.getenv("AA_CLIENT_SECRET")
)
}
\arguments{
\item{client_id}{defined by global variable or manually defined}

\item{client_secret}{defined by global variable or manually defined}
}
\description{
OAuth2 Token for Adobe Analytics
}
