% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_reports.R
\name{aa_workspace_report}
\alias{aa_workspace_report}
\title{Use a prebuilt json query to pull a ranked report}
\usage{
aa_workspace_report(
  req_body = sample,
  company_id = Sys.getenv("AA_COMPANY_ID")
)
}
\arguments{
\item{req_body}{The json string copied from workspace}
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Only runs a single dimension with as many metrics as you want.
}
