% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsfilter_map_function.R
\name{addmetricsfilters}
\alias{addmetricsfilters}
\title{Build the metrics json list for report_freeform}
\usage{
addmetricsfilters(metric, id2, filteridnumber = "0")
}
\description{
Build the metrics json list for report_freeform
}
