% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobeanalyticsr-package.R
\docType{package}
\name{adobeanalyticsr-package}
\alias{adobeanalyticsr}
\alias{adobeanalyticsr-package}
\title{\code{adobeanalyticsr} Package}
\description{
Connect to the 'Adobe Analytics' API v2.0
\url{https://github.com/AdobeDocs/analytics-2.0-apis} which powers 'Analysis
Workspace'. The package was developed with the analyst in mind, and it will
continue to be developed with the guiding principles of iterative,
repeatable, timely analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benrwoodard/adobeanalyticsr}
  \item Report bugs at \url{https://github.com/benrwoodard/adobeanalyticsr/issues}
}

}
\author{
\strong{Maintainer}: Ben Woodard \email{benrwoodard@gmail.com}

Authors:
\itemize{
  \item Tim Wilson \email{timwilson7@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Charles Gallagher \email{charlesjgallagher15@gmail.com} [contributor]
  \item Mark Edmondson [contributor]
}

}
