% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\docType{class}
\name{AdobeJwtToken}
\alias{AdobeJwtToken}
\title{Adobe JWT token response}
\format{
An R6 object
}
\description{
Includes the response object containing the bearer token as well as the
credentials used to generate the token for seamless refreshing.
}
\details{
Refreshing is disabled if the user used a custom JWT token.
}
\section{Methods}{

\itemize{
\item \code{refresh()}: refresh access token (if possible)
\item \code{validate()}: TRUE if the token is still valid, FALSE otherwise
}
}

\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AdobeJwtToken$new()}}
\item \href{#method-can_refresh}{\code{AdobeJwtToken$can_refresh()}}
\item \href{#method-refresh}{\code{AdobeJwtToken$refresh()}}
\item \href{#method-validate}{\code{AdobeJwtToken$validate()}}
\item \href{#method-clone}{\code{AdobeJwtToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdobeJwtToken$new(token, secrets)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-can_refresh"></a>}}
\if{latex}{\out{\hypertarget{method-can_refresh}{}}}
\subsection{Method \code{can_refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdobeJwtToken$can_refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-refresh}{}}}
\subsection{Method \code{refresh()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdobeJwtToken$refresh()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdobeJwtToken$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdobeJwtToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
