% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_jwt.R
\name{auth_jwt}
\alias{auth_jwt}
\title{Authenticate with the AdobeAnalytics API V2.0 using JWT}
\usage{
auth_jwt(
  config_file = Sys.getenv("AW_JWT_CONFIG_FILE"),
  key_file = Sys.getenv("AW_JWT_PRIVATE_KEY_FILE")
)
}
\arguments{
\item{config_file}{path to the JWT interface configuration JSON file.
If not provided, will attempt to read from the environment
variable \code{AW_JWT_CONFIG_FILE}}

\item{key_file}{path to the private key used to sign the JWT
If not provided, will attempt to read from the environment
variable \code{AW_JWT_PRIVATE_KEY_FILE}}
}
\value{
Auth token
}
\description{
Authenticate with the AdobeAnalytics API V2.0 using JWT
}
