% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly_report.R
\name{aw_anomaly_report}
\alias{aw_anomaly_report}
\title{Anomaly Report}
\usage{
aw_anomaly_report(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  date_range = c(Sys.Date() - 31, Sys.Date() - 1),
  metrics,
  pages = 0,
  granularity = "day",
  dateSort = "desc",
  segmentId = NA,
  anomalyDetection = TRUE
)
}
\arguments{
\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{rsid}{Adobe report number}

\item{date_range}{A two length vector of start and end Date objects (default set to show last 30 days)}

\item{metrics}{Metric to request the anomaly detection. If multiple metrics, each metric and date will have it's own row.}

\item{pages}{number of report pages}

\item{granularity}{Use either hour, day (default), week, or month}

\item{dateSort}{either by 'desc' or 'asc' order}

\item{segmentId}{Use segments to globally filter the results. Use 1 or many.}

\item{anomalyDetection}{logical statement for including anomaly. Default is TRUE}
}
\description{
Get an anomaly report for one or more metrics
}
