% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calculatedmetric_byid.R
\name{aw_get_calculatedmetric_byid}
\alias{aw_get_calculatedmetric_byid}
\title{Get a single calculated metric.}
\usage{
aw_get_calculatedmetric_byid(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  id = NA,
  locale = "en_US",
  expansion = NA
)
}
\arguments{
\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{id}{The calculated metric ID to retrieve - required}

\item{locale}{Locale. Default = en_US}

\item{expansion}{Comma-delimited list of additional calculated metric metadata fields to include on response. (reportSuiteName, ownerFullName, modified, tags, definition, categories)}
}
\description{
Retrieve one calculated metric
}
