% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{aw_get_dimensions}
\alias{aw_get_dimensions}
\title{Get list of dimensions}
\usage{
aw_get_dimensions(
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "false",
  reportable = "false",
  classifiable = "false",
  expansion = "tags,allowedForReporting,categories",
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report suite ID (RSID).  If an environment variable called \code{AW_REPORTSUITE_ID} exists
in \code{.Renviron} or elsewhere and no \code{rsid} argument is provided, then the \code{AW_REPORTSUITE_ID} value will
be used.}

\item{locale}{language - default'en_US'}

\item{segmentable}{\itemize{
\item boolean - default TRUE
}}

\item{reportable}{\itemize{
\item boolean - default TRUE
}}

\item{classifiable}{\itemize{
\item boolean - default FALSE
}}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}

\item{company_id}{Company ID. If an environment variable called \code{AW_COMPANY_ID} exists in \code{.Renviron} or
elsewhere and no \code{company_id} argument is provided, then the \code{AW_COMPANY_ID} value will be used.}
}
\description{
This will generate an extensive list of all the dimensions in the reportsuite.
}
