% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{aw_get_metrics}
\alias{aw_get_metrics}
\title{Get list of metrics}
\usage{
aw_get_metrics(
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "NULL",
  expansion = NA,
  company_id = Sys.getenv("AW_COMPANY_ID"),
  debug = FALSE
)
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{\itemize{
\item boolean - default NULL
}}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}

\item{company_id}{define globally or manually}

\item{debug}{Include the output and input of the api call in the console for debugging. Default is FALSE}
}
\description{
This will generate an extensive list of all the metrics (measures) in the reportsuite.
}
