% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reportsuites.R
\name{aw_get_reportsuites}
\alias{aw_get_reportsuites}
\title{Get list of report suites}
\usage{
aw_get_reportsuites(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  rsids = "",
  rsidContains = "",
  limit = 10,
  page = 0,
  expansion = NA,
  debug = FALSE
)
}
\arguments{
\item{company_id}{this can be either manually defined here or utilize the global definitions}

\item{rsids}{Filter list to only include suites in this RSID list (comma-delimited) Required}

\item{rsidContains}{Filter list to only include suites whose rsid contains (only one string at a time)}

\item{limit}{Number of results per page}

\item{page}{Page number (base 0 - first page is "0")}

\item{expansion}{options  c(name, parentRsid, currency, calendarType, timezoneZoneinfo)}

\item{debug}{Include the output and input of the api call in the console for debugging. Default is FALSE}
}
\description{
Get list of report suites
}
