% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_reports.R
\name{aw_workspace_report}
\alias{aw_workspace_report}
\title{Use a prebuilt json query to pull a ranked report}
\usage{
aw_workspace_report(
  req_body = "",
  company_id = Sys.getenv("AW_COMPANY_ID"),
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET")
)
}
\arguments{
\item{req_body}{The json string copied from Workspace}

\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Only runs a single dimension with as many metrics as you want.
}
