% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallout_report.R
\name{fallout_report}
\alias{fallout_report}
\title{fallout_report}
\usage{
fallout_report(
  rsid,
  date.from,
  date.to,
  metrics,
  element,
  checkpoints,
  max_attempts = 3
)
}
\arguments{
\item{rsid}{report suite id}

\item{date.from}{Start date for the report (YYYY-MM-DD)}

\item{date.to}{End date for the report (YYYY-MM-DD)}

\item{metrics}{List of metrics to include in the report}

\item{element}{Single pathed element (usually 'page')}

\item{checkpoints}{Character vector of checkpoints in the fallout path}

\item{max_attempts}{Number of API attempts to get the report before stopping (defaults to 2)
You may need to increase this for larger result sets as it may take more
attempts to get the report.}
}
\description{
A fallout report shows how visitors drop out as part of a specified path
}
\examples{
\dontrun{
df <- fallout_report("reportSuiteId", "2021-08-01", "2021-08-02",
                     metric = "pageviews",
                     element = "page",
                     checkpoints = c("Home", "Page 2"))
}

}
