% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cm_functions.R
\name{get_cm_functions}
\alias{get_cm_functions}
\title{Get Calculated Metric Functions}
\usage{
get_cm_functions(
  id = NULL,
  debug = FALSE,
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{id}{Returns details around a single calculated metric function if you specify the id.
You can obtain the desired id by not including an ID value and finding the function in the results.}

\item{debug}{Set to \code{TRUE} to publish the full JSON request(s) being sent to the API to the console when the
function is called. The default is \code{FALSE}.}

\item{company_id}{Company ID. If an environment variable called \code{AW_COMPANY_ID} exists in \code{.Renviron} or
elsewhere and no \code{company_id} argument is provided, then the \code{AW_COMPANY_ID} value will be used.
Use \code{\link[=get_me]{get_me()}} to get a list of available \code{company_id} values.}
}
\value{
Returns a full list of calculated metric functions or a speicfied function that the user can access.
}
\description{
Returns a full list of calculated metric functions that the user can access.
}
