% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_jwt.R
\name{get_jwt}
\alias{get_jwt}
\title{Create a JWT which can ben exchanged with the AdobeAnalytics API V2.0 for an
authentication token (bearer).}
\usage{
get_jwt(config, key_file)
}
\arguments{
\item{config}{list containing the JWT integration settings downloaded from the
Adobe I/O Console.}

\item{key_file}{path to the RSA private key which will be used to sign the JWT}
}
\value{
jwt A signed JWT encoded as a Base64 string
}
\description{
Create a JWT which can ben exchanged with the AdobeAnalytics API V2.0 for an
authentication token (bearer).
}
\keyword{internal}
