% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_auth.R
\name{get_jwt_token}
\alias{get_jwt_token}
\title{Get an encoded, signed JWT token}
\usage{
get_jwt_token(jwt_token = NULL, client_id, private_key, org_id, tech_id)
}
\arguments{
\item{jwt_token}{Optional, a JWT token (e.g., a cached token)}

\item{client_id}{Client ID}

\item{private_key}{File path to private key for token signature}

\item{org_id}{Organization ID from integration console}

\item{tech_id}{Technical account ID from integration console}
}
\value{
A JWT token generated by \code{\link[jose:jwt_encode]{jose::jwt_encode_sig()}}
}
\description{
Gets a JWT token
}
