% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me.R
\name{get_me}
\alias{get_me}
\title{Get all the company ids you are authorized to pull data from using the API}
\usage{
get_me(
  req_path = "discovery/me",
  client_id = Sys.getenv("AA_CLIENT_ID"),
  client_secret = Sys.getenv("AA_CLIENT_SECRET")
)
}
\arguments{
\item{req_path}{The endpoint for that particular report}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}
}
\description{
Get all the company ids you are authorized to pull data from using the API
}
\examples{

\dontrun{

get_me()

}

}
