% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{GetReport}
\usage{
get_report(
  report.id,
  interval.seconds = 10,
  max.attempts = 3,
  print.attempts = FALSE,
  format = "json",
  page = 0
)
}
\value{
Data frame
}
\description{
Internal function polls the Reporting 1.4 API for the given report id and
fetches the report data when available.
}
