% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reportsuites.R
\name{get_reportsuites}
\alias{get_reportsuites}
\title{Get list of reportsuites}
\usage{
get_reportsuites(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsids = "",
  rsidContains = "",
  limit = 10,
  page = 0,
  expansion = "name"
)
}
\arguments{
\item{company_id}{this can be either manually defined here or utilize the global definitions}

\item{rsids}{Filter list to only include suites in this RSID list (comma-delimited)}

\item{rsidContains}{Filter list to only include suites whose rsid contains rsidContains}

\item{limit}{Number of results per page}

\item{page}{Page number (base 0 - first page is "0")}

\item{expansion}{options  c(name, parentRsid, currency, calendarType, timezoneZoneinfo))}
}
\description{
Get list of reportsuites
}
