% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_api.R
\name{get_token_config}
\alias{get_token_config}
\title{Get token configuration for GET}
\usage{
get_token_config(client_id, client_secret)
}
\arguments{
\item{client_id}{Client ID}

\item{client_secret}{Client secret}
}
\value{
Config objects that can be passed to \code{httr::GET} or similar
functions (e.g. \code{httr::RETRY})
}
\description{
Returns a configuration for \code{httr::GET} for the correct token type.
}
