% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_auth.R
\name{is_jwt_expired}
\alias{is_jwt_expired}
\title{Check whether token expired (internal)}
\usage{
is_jwt_expired(token)
}
\arguments{
\item{token}{The access token to check}
}
\value{
TRUE or FALSE
}
\description{
Check whether token expired (internal)
}
