% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_timeframe}
\alias{make_timeframe}
\title{Used to format date and time}
\usage{
make_timeframe(start_date, end_date)
}
\arguments{
\item{start_date}{the first full day of data}

\item{end_date}{the last full day of data}
}
\description{
Used to format date and time
}
