% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathing_report.R
\name{pathing_report}
\alias{pathing_report}
\title{A pathing report is a report showing how often visitors go from
page A to page B to page C on the site.}
\usage{
pathing_report(
  rsid,
  date.from,
  date.to,
  metric,
  element,
  pattern,
  top = 1000,
  start = 1,
  max_attempts = 2
)
}
\arguments{
\item{rsid}{Report suite id}

\item{date.from}{Start date for the report (YYYY-MM-DD)}

\item{date.to}{End date for the report (YYYY-MM-DD)}

\item{metric}{Single metric to include in the report (usually 'pageviews')}

\item{element}{Single pathed element (usually 'page')}

\item{pattern}{Character vector of items in the path (up to 3)
use '::anything::' as a wildcard. i.e.
c("Home", "::anything::", "::anything::")
will return all paths that start with the "home" page.}

\item{top}{Number of rows to return (defaults to 1000)}

\item{start}{Row offset}

\item{max_attempts}{Number of API attempts to get the report before stopping (defaults to 2)
You may need to increase this for larger result sets as it may take more
attempts to get the report.}
}
\description{
A pathing report is a report showing how often visitors go from
page A to page B to page C on the site.
}
\examples{
\dontrun{
path_pattern <- c("Home", "::anything::", "::anyting::")
report <- pathing_report("YourReportSuiteId",
                         "2021-01-01",
                         "2021-01-31",
                         metric="pageviews",
                         element="page",
                         pattern=path_pattern)
}

}
