% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{retrieve_aw_token}
\alias{retrieve_aw_token}
\title{Retrieve a token}
\usage{
retrieve_aw_token(...)
}
\arguments{
\item{...}{Further arguments passed to auth functions}
}
\value{
A token object of type \code{response} (JWT) or \code{Token2.0} (OAuth)
}
\description{
Updates (if necessary) and returns a session token. This function first checks
for a session token, then for a cached token, and, finally, generates a
new token. The default type may be set for the session with \code{aw_auth_with()}.
}
\keyword{internal}
