% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowsdata.r
\name{rowsdata}
\alias{rowsdata}
\title{Used to join previous rows of data}
\usage{
rowsdata(it)
}
\arguments{
\item{it}{joining the itemIds and corresponding data for each new row of data}
}
\description{
Used to join previous rows of data
}
