% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_container.R
\name{seg_con}
\alias{seg_con}
\title{Create the Segment Container}
\usage{
seg_con(
  context = "hits",
  conjunction = "and",
  predicates = NULL,
  exclude = FALSE
)
}
\arguments{
\item{context}{One of hits, visits, or visitors. Also known as scope}

\item{conjunction}{This defines the relationship of the predicates.  \code{And} (default) and \code{or} are the two options.}

\item{predicates}{List of predicates created using \code{seg_pred()} function. Must wrapped in a list() function.}

\item{exclude}{Exclude the entire container}
}
\value{
a structured list of containers to be used to build the segment
}
\description{
This function combines predicates into a container.
}
