% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_delete.R
\name{seg_delete}
\alias{seg_delete}
\title{Delete A Segment}
\usage{
seg_delete(
  id = NULL,
  warn = TRUE,
  locale = "en_US",
  debug = FALSE,
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{id}{Segment ID to be deleted.}

\item{warn}{Boolean of whether or not to include a warning message.}

\item{locale}{language - default 'en_US'}

\item{debug}{Default \code{FALSE}. Set this to TRUE to see the information about the api calls as they happen.}

\item{rsid}{Adobe report suite ID (RSID).  If an environment variable called
\code{AW_REPORTSUITE_ID} exists in \code{.Renviron} or elsewhere and no \code{rsid} argument
is provided, then the \code{AW_REPORTSUITE_ID} value will be used. Use \code{\link[=aw_get_reportsuites]{aw_get_reportsuites()}}
to get a list of available \code{rsid} values.}

\item{company_id}{Company ID. If an environment variable called \code{AW_COMPANY_ID}
exists in \code{.Renviron} or elsewhere and no \code{company_id} argument is provided,
then the \code{AW_COMPANY_ID} value will be used. Use \code{\link[=get_me]{get_me()}} to get a list of
available \code{company_id} values.}
}
\value{
A string confirming the segment has or has not been deleted
}
\description{
Use this function to delete a specific segment in Adobe Analytics
}
