% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_predicate.R
\name{seg_pred}
\alias{seg_pred}
\title{Create the Segment Predicate
This function creates the simple predicate of a segment.}
\usage{
seg_pred(
  subject = "page",
  verb = "contains",
  object = "pagename",
  description = NULL,
  is_distinct = FALSE,
  attribution = "repeating",
  attribution_context = "hits",
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{subject}{This is the subject of the predicate. It is the dimension or metric id.}

\item{verb}{Choose from any of the 30 different verbs}

\item{object}{This is the object of the predicate and answer the question what or how many}

\item{description}{The internal description for the predicate if wanted}

\item{is_distinct}{Count the distinct items to compare against instead of dimension number}

\item{attribution}{Repeating (default), instance, or non-repeating instance}

\item{attribution_context}{When applying a non-repeating instance attribution model to a predicate the context for the attribution must be provided}

\item{rsid}{This is the report suite that the segment will be referenced to.}

\item{company_id}{This is the report suite that the segment will be referenced to.}
}
\value{
A structured list defining the predicate for a segment
}
\description{
Create the Segment Predicate
This function creates the simple predicate of a segment.
}
