% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_sequence.R
\name{seg_seq}
\alias{seg_seq}
\title{Create the Segment Sequence Container}
\usage{
seg_seq(
  context = "visits",
  predicates = NULL,
  sequence = "in_order",
  exclude = FALSE,
  exclude_checkpoint = NULL
)
}
\arguments{
\item{context}{One of hits, visits, or visitors. Also known as scope}

\item{predicates}{List of predicates created using \code{seg_pred()} function. Must wrapped in a list() function.}

\item{sequence}{How should the sequence of items be considered. Options: \code{in_order} (default), \code{before}, \code{after}, \code{and}, \code{or}}

\item{exclude}{Excludes the entire sequence container which will include all predicates.}

\item{exclude_checkpoint}{Which checkpoints (predicates) should be excluded. Example \code{c(1, 4)}}
}
\value{
a structured list of containers to be used to build the segment
}
\description{
This function combines predicates into a sequence container.
}
\details{
\strong{Exclude checkpoint}: Ensures the next checkpoint doesn't happen between the
preceding checkpoint and the subsequent checkpoint. If there is no subsequent
checkpoint then the excluded checkpoint must not occur at any point after
the preceding checkpoint. If there is no preceding checkpoint then the
excluded checkpoint must not have occurred at any point preceding the
subsequent checkpoint.
}
