% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{stash_token}
\alias{stash_token}
\title{Cache token}
\usage{
stash_token()
}
\value{
Invisibly, the filename of the token
}
\description{
The token is saved as \code{aw_auth.rds} in the current working directory, or
in the location pointed to by the option \code{adobeanalyticsr.auth_path}.
The \code{adobeanalyticsr.auth_name} option can be used to override the default
filename. In a pinch, this could be used to store multiple tokens.
}
