% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{token_path}
\alias{token_path}
\title{Standard location for token caching}
\usage{
token_path(...)
}
\arguments{
\item{...}{Passed to file.path, usually a filename}
}
\value{
File path
}
\description{
The default path for the token is the current working directory, but
the option \code{adobeanalyticsr.auth_path} overrides this behavior.
}
