% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_auth.R
\name{token_type}
\alias{token_type}
\title{Get type of token}
\usage{
token_type(token)
}
\arguments{
\item{token}{An \code{httr} \code{reponse} object or \code{oauth2.0_token} object}
}
\value{
Either 'oauth' or 'jwt'
}
\description{
In the future, this could become a custom object, but for now there
are too many differences between the token types, so I just check for
each class.
}
