% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_reports_freeform_metrics.R
\name{aa_breakdown_report}
\alias{aa_breakdown_report}
\title{Get a ranked report with a breakdown}
\usage{
aa_breakdown_report(date_range, metrics, dimensions, top = 50,
  metricSort = FALSE, metricFilters = NULL, dimensionSort = "asc",
  rsid = Sys.getenv("AA_RSID"))
}
\arguments{
\item{date_range}{A two length vector of start and end Date objects}

\item{metrics}{Metric to send}

\item{dimensions}{Dimension to send}

\item{rsid}{Adobe report number}

\item{num_rows}{How many rows}
}
\description{
Get a ranked report with a breakdown
}
