% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeform_report.R
\name{aa_freeform_report}
\alias{aa_freeform_report}
\title{Get a free form report}
\usage{
aa_freeform_report(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  date_range = c("2020-08-01", "2020-09-25"),
  dimensions = c("page", "lasttouchchannel", "mobiledevicetype"),
  metrics = c("visits", "visitors"),
  top = c(5),
  filterType = "breakdown",
  segmentId = NA,
  metricSort = "desc",
  include_unspecified = TRUE,
  debug = FALSE
)
}
\arguments{
\item{rsid}{Adobe report suite id number. Taken from the global environment by default if not provided.}

\item{date_range}{A two length vector of start and end Date objects}

\item{dimensions}{Dimension to send}

\item{metrics}{Metric to send}

\item{top}{How many rows. Default is set to 5. If using 'daterangeday' as the first variable you can either use only one limi number or you can add
a 0 as the first in the list of numbers. The function will then calculate how many days are included for you. This only works
if daterangeday is the first dimension listed.}

\item{filterType}{Default is 'breakdown'. This will only change if a segment is used.}

\item{segmentId}{use segments to globally filter the results. Use 1 or many.}

\item{metricSort}{Presorts the table by metrics. Values are either 'asc' or 'desc'.}

\item{include_unspecified}{TRUE is equal to "return-nones" and is set as the default}

\item{debug}{default is TRUE but set to TRUE to see the json request being sent to the Adobe API}
}
\value{
Data Frame
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Up to 7 dimensions at this time.
}
