% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeform_report.R
\name{aa_freeform_report}
\alias{aa_freeform_report}
\title{Get a free form report}
\usage{
aa_freeform_report(
  company_id = Sys.getenv("AA_COMPANY_ID"),
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  date_range = c("2020-08-01", "2020-09-25"),
  dimensions = c("page", "lasttouchchannel"),
  metrics = c("visits", "visitors"),
  top = c(5),
  filterType = "breakdown",
  metricSort = "desc"
)
}
\arguments{
\item{rsid}{Adobe report suite id number. Taken from the global environment by default if not provided.}

\item{date_range}{A two length vector of start and end Date objects}

\item{dimensions}{Dimension to send}

\item{metrics}{Metric to send}

\item{top}{How many rows. Defualt is set to 50}

\item{filterType}{Default is 'breakdown'. This will only change if a segment is used.}

\item{metricSort}{Presorts the table by metrics. Values are either 'asc' or 'desc'.}
}
\value{
Data Frame
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Up to 7 dimensions at this time.
}
