% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{aa_get_dimensions}
\alias{aa_get_dimensions}
\title{Get list of dimensions}
\usage{
aa_get_dimensions(rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  locale = "en_US", segmentable = "false", reportable = "false",
  classifiable = "false",
  expansion = "tags,allowedForReporting,categories",
  company_id = Sys.getenv("AA_COMPANY_ID"))
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{- boolean - default TRUE}

\item{reportable}{= boolean - default TRUE}

\item{classifiable}{= boolean - default FALSE}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}
}
\description{
Get list of dimensions
}
