% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{aa_get_metrics}
\alias{aa_get_metrics}
\title{Get list of dimensions}
\usage{
aa_get_metrics(
  rsid = Sys.getenv("AA_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "NULL",
  expansion = "tags, allowedForReporting, categories",
  company_id = Sys.getenv("AA_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{- boolean - default NULL}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}

\item{company_id}{define globally or manually}
}
\description{
Get list of dimensions
}
