% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_map_function.R
\name{addmetrics}
\alias{addmetrics}
\title{Build the metrics json list for report_freeform}
\usage{
addmetrics(metric, id2, metricSort)
}
\description{
Build the metrics json list for report_freeform
}
