% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adobeanalyticsr-package.R
\docType{package}
\name{adobeanalyticsr-package}
\alias{adobeanalyticsr}
\alias{adobeanalyticsr-package}
\title{\code{adobeanalyticsr} Package}
\description{
Description of \code{adobeanalyticsr}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/benrwoodard/adobeanalyticsr}
  \item Report bugs at \url{https://github.com/benrwoodard/adobeanalyticsr/issues}
}

}
\author{
\strong{Maintainer}: Ben Woodard \email{benrwoodard@gmail.com}

Authors:
\itemize{
  \item Tim Wilson \email{timwilson7@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Mark Edmondson [contributor]
}

}
