% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{aw_auth_with}
\alias{aw_auth_with}
\title{Specify the type of auth for the session}
\usage{
aw_auth_with(type)
}
\arguments{
\item{type}{Either "oauth" or "jwt"}
}
\value{
\code{type}, invisibly
}
\description{
Specify the type of auth for the session
}
