% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeform_report.R
\name{aw_freeform_report}
\alias{aw_freeform_report}
\title{Get a freeform report}
\usage{
aw_freeform_report(
  company_id = Sys.getenv("AW_COMPANY_ID"),
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  date_range = c(Sys.Date() - 30, Sys.Date() - 1),
  dimensions = c("page", "lasttouchchannel", "mobiledevicetype"),
  metrics = c("visits", "visitors"),
  top = c(5),
  page = 0,
  filterType = "breakdown",
  segmentId = NA,
  metricSort = "desc",
  include_unspecified = TRUE,
  search = NA,
  prettynames = FALSE,
  debug = FALSE
)
}
\arguments{
\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{rsid}{Adobe report suite id number. Taken from the global environment by default if not provided.}

\item{date_range}{A two length vector of start and end Date objects}

\item{dimensions}{A character vector of dimensions. There is a limit of 20 dimension breakdowns at this time.}

\item{metrics}{A character vector of metrics.}

\item{top}{The top number of dimensions you want to pull. Default is set to 5. If using 'daterange...' as the first variable you can either use only one number item (top = 5) or you can add a 0 as the top item in the list of numbers (top = c(0, 20)). The function will then calculate how many hours, minutes, days, weeks, months, years, quarters, and years should be requested. Any dimensions in the "daterange..." can be liste das '0' and the function will calculated the number of items needed.}

\item{page}{Used in combination with 'top' to return the next page of results. Uses 0 based numbering. i.e. Top 50000 + page 1 will return the top 50,000 items starting at 50,001.}

\item{filterType}{Currently, Only 'breakdown' is supported but future versions should include 'segment' breakdown capability.}

\item{segmentId}{use segments to globally filter the results. Use 1 or many.}

\item{metricSort}{Presorts the table by metrics. Values are either 'asc' or 'desc'.}

\item{include_unspecified}{TRUE is equal to "return-nones" and is set as the default}

\item{search}{The structure of a search string is specific and can be compound using "AND". Each dimension can be filtered using a unique item in a character vector item.}

\item{prettynames}{Boolean for whether the results should have the name field value for easier understanding.}

\item{debug}{default is FALSE but set to TRUE to see the json request being sent to the Adobe API}
}
\value{
Data Frame
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Up to 15 dimensions at this time.
}
