% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{aw_get_dimensions}
\alias{aw_get_dimensions}
\title{Get list of dimensions}
\usage{
aw_get_dimensions(
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = FALSE,
  reportable = FALSE,
  classifiable = FALSE,
  expansion = NA,
  debug = FALSE,
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report number - required}

\item{locale}{language - default'en_US'}

\item{segmentable}{\itemize{
\item boolean - default FALSE
}}

\item{reportable}{\itemize{
\item boolean - default FALSE
}}

\item{classifiable}{\itemize{
\item boolean - default FALSE
}}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories')}

\item{debug}{Include the output and input of the api call in the console for debugging. Default is FALSE}

\item{company_id}{this can be either manually defined here or utilize the global definitions}
}
\description{
This will generate an extensive list of all the dimensions in the reportsuite.
}
