% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{aw_get_dimensions}
\alias{aw_get_dimensions}
\title{Get list of dimensions}
\usage{
aw_get_dimensions(
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "false",
  reportable = "false",
  classifiable = "false",
  expansion = "tags,allowedForReporting,categories",
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{- boolean - default TRUE}

\item{reportable}{- boolean - default TRUE}

\item{classifiable}{- boolean - default FALSE}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}

\item{company_id}{this can be either manually defined here or utilize the global definitions}
}
\description{
This will generate an extensive list of all the dimensions in the reportsuite.
}
