% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{aw_get_metrics}
\alias{aw_get_metrics}
\title{Get list of metrics}
\usage{
aw_get_metrics(
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  locale = "en_US",
  segmentable = "NULL",
  expansion = "tags, allowedForReporting, categories",
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{rsid}{Adobe report number}

\item{locale}{language - default'en_US'}

\item{segmentable}{- boolean - default NULL}

\item{expansion}{options  c('tags', 'allowedForReporting', 'categories'))}

\item{company_id}{define globally or manually}
}
\description{
This will generate an extensive list of all the metrics (measures) in the reportsuite.
}
