% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace_reports.R
\name{aw_workspace_report}
\alias{aw_workspace_report}
\title{Use a prebuilt json query to pull a ranked report}
\usage{
aw_workspace_report(
  req_body = sample,
  company_id = Sys.getenv("AW_COMPANY_ID"),
  debug = FALSE
)
}
\arguments{
\item{req_body}{The json string copied from Workspace}

\item{company_id}{Company Id.  Taken from the global environment by default if not provided.}

\item{debug}{default is FALSE  but set to TRUE to see the json request being sent to the Adobe API}
}
\description{
Organizes the arguments into a json string and then structures the data after the internal function makes
the api call. Only runs a single dimension with as many metrics as you want.
}
