% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me.R
\name{get_me}
\alias{get_me}
\title{Get Company Ids}
\usage{
get_me(req_path = "discovery/me")
}
\arguments{
\item{req_path}{The endpoint for that particular report}
}
\value{
A data frame of company ids and company names
}
\description{
This function will quickly pull the list of company ids that you have access to.
}
\examples{
\dontrun{
get_me()
}
}
