% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me.R
\name{get_me}
\alias{get_me}
\title{Get Company Ids}
\usage{
get_me(
  req_path = "discovery/me",
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET")
)
}
\arguments{
\item{req_path}{The endpoint for that particular report}

\item{client_id}{Set in environment args, or pass directly here}

\item{client_secret}{Set in environment args, or pass directly here}
}
\description{
This function will quickly pull the list of company ids that you have access to.
}
\examples{
\dontrun{
get_me()
}

}
