% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_timeframe}
\alias{make_timeframe}
\title{make_timeframe function for formating date and time}
\usage{
make_timeframe(start_date, end_date)
}
\description{
make_timeframe function for formating date and time
}
