% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{retrieve_aw_token}
\alias{retrieve_aw_token}
\title{Retrieve a token}
\usage{
retrieve_aw_token(...)
}
\arguments{
\item{...}{Further arguments passed to auth functions}
}
\value{
A token object of type \code{response} (JWT) or \code{Token2.0} (OAuth)
}
\description{
Updates (if necessary) and returns session token. First checks for a session
token, then a cached token, then generates a new token. The default type may
be set for the session with \code{aw_auth_with}.
}
\keyword{internal}
