% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_build.R
\name{seg_build}
\alias{seg_build}
\title{Build the segment in Adobe Analytics}
\usage{
seg_build(
  name = "this is the name",
  description = "this is the description",
  containers = NULL,
  predicates = NULL,
  sequences = NULL,
  context = "hits",
  conjunction = "and",
  sequence = "in_order",
  sequence_context = "hits",
  exclude = FALSE,
  version = list(1, 0, 0),
  debug = FALSE,
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID"),
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET")
)
}
\arguments{
\item{name}{This is the name of the new segment}

\item{description}{This is the description of the segment}

\item{containers}{List of the container(s) that make up the segment}

\item{predicates}{List of the predicate(s) to create a segment}

\item{sequences}{List of the predicates and sequence containers that are combined to make a segment}

\item{context}{This defines the main container context, either hits, visits (Default), or visitors}

\item{conjunction}{This will tell how the different containers and predicates should be compared. Use either 'and' or 'or' at this time. Sequential 'then' will be added in a future update.}

\item{sequence}{Used to define if the segment should be 'in_order' (default), 'after', or 'before' the sequence of events}

\item{sequence_context}{Used to define the sequential items context which should be below the container context. ex. if container context is visitors then the sequence_context should be visits or hits}

\item{exclude}{Excludes the main container which will include all predicates. Only used when the predicate arguments are used.}

\item{version}{This is the default version. Only used if updating an existing segment. Not to be edited at this time.}

\item{debug}{This enables the api call information to show in the console for help with debugging issues. default is FALSE}

\item{rsid}{This is the report suite that the segment will be referenced to.}

\item{company_id}{This is the report suite that the segment will be referenced to.}

\item{client_id}{This is the report suite that the segment will be referenced to.}

\item{client_secret}{This is the report suite that the segment will be referenced to.}
}
\value{
If the segment validates it will return a data frame of the new created segment id and other basic elements. If it returns and error then the error
response will be returned to help understand what needs to be corrected.
}
\description{
This function combines predicates and/or containers and then makes the post call to create the segment in Adobe Analytics.
}
