% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_build.R
\name{seg_build}
\alias{seg_build}
\title{Build the Segment}
\usage{
seg_build(
  name = "this is the name",
  description = "this is the description",
  containers = NULL,
  predicates = NULL,
  context = "hits",
  conjunction = "and",
  exclude = FALSE,
  version = list(1, 0, 0),
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID"),
  client_id = Sys.getenv("AW_CLIENT_ID"),
  client_secret = Sys.getenv("AW_CLIENT_SECRET")
)
}
\arguments{
\item{name}{This is the name of the new segment}

\item{description}{This is the description of the segment}

\item{containers}{List of the container(s) that make up the segment}

\item{predicates}{List of the predicate(s) to create a segment}

\item{context}{This defines the main container context, either hits, visits (Default), or visitors}

\item{conjunction}{This will tell how the different containers and predicates should be compared. Use either 'and' or 'or' at this time. Sequential 'then' will be added in a future update.}

\item{exclude}{excludes the main container which will include all predicates. Only used when the predicate arguments are used.}

\item{version}{This is the default version. Only used if updating an existing segment. Not to be edited at this time.}

\item{rsid}{This is the report suite that the segment will be referenced to.}

\item{company_id}{This is the report suite that the segment will be referenced to.}

\item{client_id}{This is the report suite that the segment will be referenced to.}

\item{client_secret}{This is the report suite that the segment will be referenced to.}
}
\value{
A data frame of the new created segment id and other meta elements
}
\description{
This function combines predicates into a container.
}
