% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_predicate.R
\name{seg_pred}
\alias{seg_pred}
\title{Create the Segment Predicate
This function creates the simple predicate of a segment.}
\usage{
seg_pred(
  subject = "page",
  verb = "contains",
  object = "pagename",
  is_distinct = FALSE,
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{subject}{This is the subject of the predicate. It is the dimension or metric id.}

\item{verb}{Choose from any of the 30 different verbs}

\item{object}{this is the object of the predicate and answer the question what or how many}

\item{is_distinct}{count the distinct items to compare against instead of dimension number}

\item{rsid}{This is the report suite that the segment will be referenced to.}

\item{company_id}{This is the report suite that the segment will be referenced to.}
}
\value{
A structured list defining the predicate for a segment
}
\description{
Create the Segment Predicate
This function creates the simple predicate of a segment.
}
