% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_predicate.R
\name{seg_pred}
\alias{seg_pred}
\title{Create the Segment Predicate}
\usage{
seg_pred(
  subject = "page",
  verb = "contains",
  object = "pagename",
  description = NULL,
  is_distinct = FALSE,
  attribution = "repeating",
  attribution_context = "visitors",
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{subject}{This is the subject of the predicate. It is the dimension or metric id.}

\item{verb}{Choose from any of the 30 different verbs}

\item{object}{This is the object of the predicate and answer the question what or how many}

\item{description}{The internal description for the predicate if wanted}

\item{is_distinct}{Count the distinct items to compare against instead of dimension number}

\item{attribution}{Define the type of attribution. Either \code{repeating} (default), \code{instance}, or \code{nonrepeating}}

\item{attribution_context}{When applying a non-repeating instance attribution model to a predicate the context for the attribution must be \code{visitors} (default) or \code{visits}}

\item{rsid}{Adobe report suite ID (RSID).  If an environment variable called \code{AW_REPORTSUITE_ID} exists
in \code{.Renviron} or elsewhere and no \code{rsid} argument is provided, then the \code{AW_REPORTSUITE_ID} value will
be used. Use \code{\link[=aw_get_reportsuites]{aw_get_reportsuites()}} to get a list of available \code{rsid} values.}

\item{company_id}{Company ID. If an environment variable called \code{AW_COMPANY_ID} exists in \code{.Renviron} or
elsewhere and no \code{company_id} argument is provided, then the \code{AW_COMPANY_ID} value will be used.
Use \code{\link[=get_me]{get_me()}} to get a list of available \code{company_id} values.}
}
\value{
A structured list defining the predicate for a segment
}
\description{
This function creates the simple predicate of a segment.
}
