% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_predicate.R
\name{seg_pred}
\alias{seg_pred}
\title{Create the Segment Predicate}
\usage{
seg_pred(
  subject = "page",
  verb = "contains",
  object = "pagename",
  description = NULL,
  is_distinct = FALSE,
  attribution = "repeating",
  attribution_context = "visitors",
  rsid = Sys.getenv("AW_REPORTSUITE_ID"),
  company_id = Sys.getenv("AW_COMPANY_ID")
)
}
\arguments{
\item{subject}{This is the subject of the predicate. It is the dimension or metric id.}

\item{verb}{Choose from any of the 30 different verbs. Use the \code{seg_verbs()} function to see all available verbs along with the descriptions.}

\item{object}{This is the object of the predicate and answers the question \code{what} or \verb{how many}}

\item{description}{The internal description for the predicate. (optional) This will not show in the UI but could be very helpful when using the API.}

\item{is_distinct}{This will segment on a distinct count of items within a dimension. Examples: “Visitors who viewed more than 5 distinct products,” or “Visits where more than 5 distinct pages were seen.”}

\item{attribution}{Define the type of attribution. Either \code{repeating} (default), \code{instance}, or \code{nonrepeating}. See Details for more information.}

\item{attribution_context}{When applying a non-repeating instance attribution model to a predicate the context for the attribution must be \code{visitors} (default) or \code{visits}}

\item{rsid}{Adobe report suite ID (RSID).  If an environment variable called \code{AW_REPORTSUITE_ID} exists
in \code{.Renviron} or elsewhere and no \code{rsid} argument is provided, then the \code{AW_REPORTSUITE_ID} value will
be used. Use \code{\link[=aw_get_reportsuites]{aw_get_reportsuites()}} to get a list of available \code{rsid} values.}

\item{company_id}{Company ID. If an environment variable called \code{AW_COMPANY_ID} exists in \code{.Renviron} or
elsewhere and no \code{company_id} argument is provided, then the \code{AW_COMPANY_ID} value will be used.
Use \code{\link[=get_me]{get_me()}} to get a list of available \code{company_id} values.}
}
\value{
A structured list defining the predicate for a segment
}
\description{
This function creates the simple predicate of a segment.
}
\details{
\strong{Attribution Models}

Available for dimensions only, these models determine what values in a dimension to segment for. Dimension models are particularly useful in sequential segmentation.
\emph{repeating} (default): Includes instances and persisted values for the dimension.
\emph{instance}: Includes instances for the dimension.
\emph{nonrepeating} instance: Includes unique instances (non-repeating) for the dimension. This is the model applied in Flow when repeat instances are excluded.
}
