% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_then.R
\name{seg_then}
\alias{seg_then}
\title{Create the segment sequence then object}
\usage{
seg_then(limit = "within", count = 1, unit = "week")
}
\arguments{
\item{limit}{The limitation of the restriction. Either \code{within} (default) or \code{after}}

\item{count}{How many of the units should be used}

\item{unit}{A unit of time. Valid values are \code{minute}, \code{hour}, \code{day}, \code{week} (default), \code{month}, \code{quarter}, or \code{year}.}
}
\value{
a structured list of time restrictions to be used to build the sequential segment
}
\description{
This function creates a then list object which restricts the time constraint of a segment to be added to a sequence segment.
}
\details{
\strong{Using within and after in the same time restriction:}
Time restrictions can only be combined using 'within' first before 'after'.
The function will automatically align these to be in the correct list item order.
}
