% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_verbs_list.R
\name{seg_verbs}
\alias{seg_verbs}
\title{Retrieve a list of verbs for segments}
\usage{
seg_verbs()
}
\value{
A list of all available verbs
}
\description{
This function generates a list of linking verbs that can be used to build segments
}
