% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aw_auth.R
\name{token_path}
\alias{token_path}
\title{Standard location for token caching}
\usage{
token_path(...)
}
\arguments{
\item{...}{Passed to file.path, usually a filename}
}
\value{
File path
}
\description{
Standard location for token caching
}
